refactor: move `host` to src root
